Add validation localization. Add other functions.

+++ src/main/java/by/zinkov/victor/command/impl/ActivateCommand.java
@@ -1,6 +1,7 @@
 package by.zinkov.victor.command.impl;
 
 import by.zinkov.victor.command.Command;
+import by.zinkov.victor.command.Page;
 import by.zinkov.victor.command.Router;
 import by.zinkov.victor.command.CommandException;
 import by.zinkov.victor.domain.RegistrationKey;
@@ -11,36 +12,50 @@ import by.zinkov.victor.service.UserService;
 import by.zinkov.victor.service.ServiceException;
 import by.zinkov.victor.service.impl.RegistrationKeyServiceImpl;
 import by.zinkov.victor.service.impl.UserServiceImpl;
+import by.zinkov.victor.validation.Validator;
+import by.zinkov.victor.validation.ValidatorFactory;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import java.util.Map;
 
-public class ActivateCommand implements Command {
+public class ActivateCommand extends Command {
+    private static final Logger LOGGER = LogManager.getLogger(ActivateCommand.class);
     private static final String USER_ID = "user_id";
     private static final String ACTIVATE_STRING = "value";
     private static final String USER_ATTRIBUTE = "user";
-    private static final Logger LOGGER = LogManager.getLogger(ActivateCommand.class);
+    private static final String ERRORS = "errors";
+
 
     @Override
     public Router execute(HttpServletRequest request) throws CommandException {
         Router router = new Router();
         router.setType(Router.Type.REDIRECT);
 
-        Integer id = Integer.valueOf(request.getParameter(USER_ID));
-        String value = request.getParameter(ACTIVATE_STRING);
-
+        Map<String,String> parameters = super.readParameters(request);
+        ValidatorFactory validatorFactory = ValidatorFactory.getInstance();
+        Validator validator = validatorFactory.getActivateCommandValidator();
         RegistrationKeyService service = new RegistrationKeyServiceImpl();
         try {
+            Map<String,String> errorsMap =  validator.validate(parameters);
+            if(errorsMap.size() != 0){
+                router.setRoute(Page.INDEX.getRout());
+                router.setType(Router.Type.FORWARD);
+                request.setAttribute(ERRORS,errorsMap);
+                return router;
+            }
+
+            Integer id = Integer.valueOf(parameters.get(USER_ID));
+            String value = parameters.get(ACTIVATE_STRING);
+
             RegistrationKey key = service.getById(id);
             if(key.getKey().equals(value)){
                 UserService userService = new UserServiceImpl();
                 UserDto userDto  = userService.getByPK(id);
-                userService.setNewStatus(userDto.getId(),"ACTIVE");
+                userService.setNewStatus(userDto.getId(),UserStatus.ACTIVE.toString());
                 userDto.setUserStatus(UserStatus.ACTIVE);
-
-
                 service.remove(key);
                 HttpSession session = request.getSession();
                 session.setAttribute(USER_ATTRIBUTE, userDto);
@@ -51,7 +66,6 @@ public class ActivateCommand implements Command {
         } catch (ServiceException e) {
             LOGGER.error(e);
             router.setRoute(Router.INDEX_ROUT);
-
         }
         return router;
     }
